third_party/blink/web_tests/platform/generic/fast/spatial-navigation/snav-zero-margin-content-expected.txt

test1 test2
PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
